% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syrup-package.R
\docType{package}
\name{syrup-package}
\alias{syrup-package}
\title{syrup: Measure Memory Usage for Parallel R Code}
\description{
Coarsely measures memory usage of R code run in parallel by regularly taking snapshots of calls to the system command ps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/syrup}
  \item \url{https://simonpcouch.github.io/syrup/}
  \item Report bugs at \url{https://github.com/simonpcouch/syrup/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
