% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syrup-package.R
\docType{package}
\name{syrup-package}
\alias{syrup-package}
\title{syrup: Measure Memory and CPU Usage for Parallel R Code}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Measures memory and CPU usage of R code by regularly taking snapshots of calls to the system command 'ps'. The package provides an entry point (albeit coarse) to profile usage of system resources by R code run in parallel.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/syrup}
  \item \url{https://simonpcouch.github.io/syrup/}
  \item Report bugs at \url{https://github.com/simonpcouch/syrup/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
